requirements.txt
faiss-cpu
fastapi
uvicorn
sentence-transformers
pydantic
requests
PyPDF2
python-dotenv
google-generativeai
